feat(urls): support relative CSS / template URLs in components
Example:
When defining a component called 'foo' in file
'/myproject/has/many/folders/foo.dart', prior to the change you would
do so thusly:
@Component(
   selector: 'foo',
   templateUrl: '/myproject/has/many/folders/foo.html',
   cssUrl: '/myproject/has/many/folders/foo.css'
)

The above is still valid, but now can also be defined rather more
succinctly as:
@Component(
   selector: 'foo',
   templateUrl: 'foo.html',
   cssUrl: 'foo.css'
)

A full table of what URLs will be changed/unchanged is below:
You Say:				Transformer replies:
'packages/foo/bar.html'			'packages/foo/bar.html'
'foo.html'				'packages/foo/bar.html'
'./foo.html'				'packages/foo/bar.html'
'/foo.html'				'/foo.html'
'http://google.com/foo.html'		'http://google.com/foo.html'

*Note  that as  shown any absolute paths you define will not be
bothered by the transformer, so if you have a templateUrl living at
<root>/foo.html, you can still reach it with '/foo.html'!

This feature is defaulted to an "off" state through the addition
of the ResourceResolverConfig class.  To turn the feature on, bind
a new ResourceResolverConfig in your module:

module.bind(ResourceResolverConfig, toValue:
	new ResourceResolverConfig(useRelativeUrls: true));

*Testing*:
- e2e transformer tests can now be done on the sample application found
  in the test_transformers folder

+++ scripts/run-e2e-test.sh
@@ -1,43 +1,81 @@
 #!/bin/bash
 
 # Run E2E / Protractor tests.
+#
+# Usage (from the angular.dart root folder):
+#
+#   # 1. Run all e2e tests.
+#   ./scripts/run-e2e-test.sh
+#   ./scripts/run-e2e-test.sh example transformers # same as previous
 
-set -e -o pipefail
+#   # 2. E2E test for the examples.
+#   ./scripts/run-e2e-test.sh example
+#
+#   # 3. E2E test for transformers.
+#   ./scripts/run-e2e-test.sh transformers
+
+set -eE -o pipefail
 
 . $(dirname $0)/env.sh
 
-SIGNALS=(ERR HUP INT QUIT PIPE TERM)
+SIGNALS=(ERR HUP INT QUIT PIPE TERM EXIT)
+export SIGNALS
+
 
 _onSignal() {
-  EXIT_CODE=$?
-  # Kill all child processes (running servers.)
-  kill 0
+  trap - "${SIGNALS[@]}"  # disable signals so we don't capture them again.
+  # Kill all child processes (running servers.)  On Travis, kill 0 hangs for some
+  # reason but I can't get it to repro on any other machine.  We don't need
+  # cleanup on Travis anyway since the VM will be shutdown.
+  if [[ -z $TRAVIS ]]; then 
+    kill 0
+  fi
   # Need to explicitly kill ourselves to let the caller know we died from a
   # signal.  Ref: http://www.cons.org/cracauer/sigint.html
   sig=$1
-  trap - "${SIGNALS[@]}"  # disable signals so we don't capture them again.
-  if [[ "$sig" == "ERR" ]]; then
-    exit $EXIT_CODE
-  else
+  if [[ "$sig" != "ERR" ]]; then
     kill -$sig $$
   fi
+  exit $EXIT_CODE
 }
+export -f _onSignal
 
-for s in "${SIGNALS[@]}" ; do
-  trap "_onSignal $s" $s
-done
+_initSignals() {
+  for s in "${SIGNALS[@]}" ; do
+    trap "EXIT_CODE=$? ; _onSignal $s" $s
+  done
+}
+export -f _initSignals
+
+_initSignals
+
+
+PLATFORM="$(uname -s)"
+
+case "$PLATFORM" in
+  (Darwin)
+    parallelize_shell_cmd() {
+      xargs -n 1 -P 4 -I ARG bash -c "set -vx; _initSignals; $*"
+    }
+    CHROMEDRIVER_ZIP="chromedriver_mac32.zip"
+    ;;
+  (Linux)
+    parallelize_shell_cmd() {
+      xargs -d ' ' -n 1 -P 4 -I ARG bash -c "set -vx; _initSignals; $*"
+    }
+    CHROMEDRIVER_ZIP="chromedriver_linux64.zip"
+    ;;
+  (*)
+    echo Unsupported platform $PLATFORM.  Exiting ... >&2
+    exit 3
+    ;;
+esac
 
 
 install_deps() {(
   SELENIUM_VER="2.42"
   SELENIUM_ZIP="selenium-server-standalone-$SELENIUM_VER.0.jar"
   CHROMEDRIVER_VER="2.10"
-  # chromedriver
-  case "$(uname -s)" in
-    (Darwin) CHROMEDRIVER_ZIP="chromedriver_mac32.zip" ;;
-    (Linux)  CHROMEDRIVER_ZIP="chromedriver_linux64.zip" ;;
-    (*) echo Unsupported OS >&2; exit 2 ;;
-  esac
   mkdir -p e2e_bin && cd e2e_bin
   if [[ ! -e "$SELENIUM_ZIP" ]]; then
     curl -O "http://selenium-release.storage.googleapis.com/$SELENIUM_VER/$SELENIUM_ZIP"
@@ -49,19 +87,26 @@ install_deps() {(
 )}
 
 
-start_servers() {
-  # Run examples.
-  ( 
-    cd example
+function start_all_servers() {(
+  local TEST_TYPES=("$@")
+
+  build_app_and_serve() {(
+    local TEST_TYPE=$1
+    local APP_ROOT_VAR=TEST_${TEST_TYPE}_APP_ROOT
+    cd ${!APP_ROOT_VAR}
     pub install
     pub build
     rsync -rl --exclude packages web/ build/web/
     rm -rf build/web/packages
     ln -s $PWD/packages build/web/packages
-  )
-  PORT=28000
-  (cd example/build/web && python -m SimpleHTTPServer $PORT) >/dev/null 2>&1 &
-  export NGDART_EXAMPLE_BASEURL=http://127.0.0.1:$PORT
+
+    # now serve it.
+    local PORT_VAR=TEST_${TEST_TYPE}_PORT
+    (cd build/web && python -m SimpleHTTPServer ${!PORT_VAR}) >/dev/null 2>&1 &
+  )}
+  export -f build_app_and_serve
+
+  echo ${TEST_TYPES[@]} | parallelize_shell_cmd 'build_app_and_serve ARG'
 
   # Allow chromedriver to be found on the system path.
   export PATH=$PATH:$PWD/e2e_bin
@@ -70,11 +115,47 @@ start_servers() {
   java -jar ./e2e_bin/selenium-server-standalone-2.42.0.jar >/dev/null 2>&1 &
 
   sleep 4 # wait for selenium startup
-}
+)}
+
+
+# Config
+
+export TEST_EXAMPLE_PORT=28000
+export TEST_EXAMPLE_APP_ROOT=example
+export TEST_EXAMPLE_CONF=test_e2e/examplesConf.js
+export TEST_EXAMPLE_BASEURL=http://127.0.0.1:$TEST_EXAMPLE_PORT
+
+export TEST_TRANSFORMERS_PORT=28100
+export TEST_TRANSFORMERS_APP_ROOT=test_transformers
+export TEST_TRANSFORMERS_CONF=test_transformers/transformersE2eConf.js
+export TEST_TRANSFORMERS_BASEURL=http://127.0.0.1:$TEST_TRANSFORMERS_PORT
 
 
 # Main
+
+if [[ ${#@} == "0" ]]; then
+  set -- example transformers
+fi
+TEST_TYPES=($(echo "$@" | tr '[:lower:]' '[:upper:]'))
+
 install_deps
-start_servers
+start_all_servers ${TEST_TYPES[@]}
+
+echo pstree after starting servers
+if which pstree ; then
+  pstree $$
+fi
+
+
+# TODO: REFACTOR.
 (cd test_e2e && pub install)
-./node_modules/.bin/protractor_dart test_e2e/examplesConf.js
+(cd test_transformers && pub install)
+
+run_protractor() {(
+  local TEST_TYPE=$1
+  local SPEC_FILE_VAR=TEST_${TEST_TYPE}_CONF
+  ./node_modules/.bin/protractor_dart ${!SPEC_FILE_VAR}
+)}
+export -f run_protractor
+
+echo ${TEST_TYPES[@]} | parallelize_shell_cmd 'run_protractor ARG'
